+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/jni/JavaCallback.kt
@@ -6,6 +6,7 @@ import com.facebook.react.bridge.WritableNativeMap
 import expo.modules.core.interfaces.DoNotStrip
 import expo.modules.kotlin.exception.UnexpectedException
 import expo.modules.kotlin.logger
+import expo.modules.kotlin.sharedobjects.SharedObject
 import expo.modules.kotlin.sharedobjects.SharedRef
 
 @Suppress("KotlinJniMissingFunction")
@@ -24,7 +25,7 @@ class JavaCallback @DoNotStrip internal constructor(@DoNotStrip private val mHyb
       is String -> invokeNative(result)
       is WritableNativeArray -> invokeNative(result)
       is WritableNativeMap -> invokeNative(result)
-      is SharedRef<*> -> invokeNative(result)
+      is SharedObject -> invokeNative(result)
       else -> throw UnexpectedException("Unknown type: ${result.javaClass}")
     }
   }
@@ -65,7 +66,7 @@ class JavaCallback @DoNotStrip internal constructor(@DoNotStrip private val mHyb
   private external fun invokeNative(result: String)
   private external fun invokeNative(result: WritableNativeArray)
   private external fun invokeNative(result: WritableNativeMap)
-  private external fun invokeNative(result: SharedRef<*>)
+  private external fun invokeNative(result: SharedObject)
   private external fun invokeNative(code: String, errorMessage: String)
 
   private inline fun checkIfValid(body: () -> Unit) {